Module hyper_system_resolver::background

Background resolver.

This module contains utilities to aid with implementing resolvers that run in the background in a tokio thread pool.

You can provide a custom Resolve trait implementation that uses either tokio::spawn, tokio::task::spawn_blocking or anything else that returns tokio::task::JoinHandle.

Structs

Resolver

A hyper-compatible resolver implementation. Delegates the actual resolution logic to the generic parameter T.

ResolverFuture

The opaque resolver future.

Traits

Resolve

Resolve the name in the background.